Try with ImmersiveModalContent with its own style file
